Adding app files
 and restructured revision of recent merges
