--- MachineLearning/LCMS/df_lcms_ee.csv
@@ -1,2 +1,0 @@
-Name,Base model Balanced Accuracy,Grid model Balanced Accuracy,Improvement
-Easy Ensemble AdaBoost,0.8524247491638797,0.8524247491638797,0.00%
