--- MachineLearning/LCMS/df_lcms_xgb.csv
@@ -1,2 +1,0 @@
-Name,Base model Balanced Accuracy,Grid model Balanced Accuracy,Improvement
-XGBoost,0.8658026755852843,0.8793896321070234,1.57%
